Move delta report actions and selector to report modules
Move the delta report actions and delta report selector to the more
specific report modules.
