+++ gsa/src/web/store/entities/reports.js
@@ -16,10 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-import {isDefined} from 'gmp/utils/identity';
-
 import {
-  createEntityActions,
   createEntitiesActions,
   createLoadAllEntities,
   createLoadEntities,
@@ -63,70 +60,11 @@ const reducer = (state = initialState, action) => {
       return state;
   }
 };
-const entityType = 'deltaReport';
 
-const deltaIdentifier = (id, deltaId) => `${id}+${deltaId}`;
-
-class DeltaSelector {
-  constructor(state = {}) {
-    this.state = state;
-  }
-
-  isLoading(id, deltaId) {
-    return isDefined(this.state.isLoading)
-      ? !!this.state.isLoading[deltaIdentifier(id, deltaId)]
-      : false;
-  }
-
-  getError(id, deltaId) {
-    return isDefined(this.state.errors)
-      ? this.state.errors[deltaIdentifier(id, deltaId)]
-      : undefined;
-  }
-
-  getEntity(id, deltaId) {
-    return isDefined(this.state.byId)
-      ? this.state.byId[deltaIdentifier(id, deltaId)]
-      : undefined;
-  }
-}
-
-const deltaReducer = createReducer(entityType);
-const deltaSelector = rootState =>
-  new DeltaSelector(rootState.entities[entityType]);
-
-const deltaEntityActions = createEntityActions(entityType);
-
-const loadDeltaReport = gmp => (id, deltaId, filter) => (
-  dispatch,
-  getState,
-) => {
-  const rootState = getState();
-  const state = deltaSelector(rootState);
-
-  if (state.isLoading(id, deltaId)) {
-    // we are already loading data
-    return Promise.resolve();
-  }
-
-  const identifier = deltaIdentifier(id, deltaId);
-
-  dispatch(deltaEntityActions.request(identifier));
-
-  return gmp.report
-    .getDelta({id}, {id: deltaId}, {filter})
-    .then(
-      response =>
-        dispatch(deltaEntityActions.success(identifier, response.data)),
-      error => dispatch(deltaEntityActions.error(identifier, error)),
-    );
-};
+const deltaReducer = createReducer('deltaReport');
 
 export {
-  deltaEntityActions,
   deltaReducer,
-  deltaSelector,
-  loadDeltaReport,
   loadAllEntities,
   loadEntities,
   reducer,
